+++ src/main/java/io/hhplus/tdd/common/api/advice/ApiControllerAdvice.java
@@ -1,9 +1,11 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.common.api.advice;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import io.hhplus.tdd.common.api.support.error.ErrorResponse;
 
 @RestControllerAdvice
 class ApiControllerAdvice extends ResponseEntityExceptionHandler {
