+++ src/main/java/io/hhplus/tdd/common/api/support/error/ErrorResponse.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.common.api.support.error;
 
 public record ErrorResponse(
         String code,
